+++ src/main/java/com/sns/api/auth/controller/AuthController.java
@@ -1,10 +1,10 @@
 package com.sns.api.auth.controller;
 
-import com.sns.api.auth.domain.dto.LoginRequestDto;
-import com.sns.api.auth.domain.dto.SignupRequestDto;
+import com.sns.api.auth.domain.dto.request.LoginRequestDto;
+import com.sns.api.auth.domain.dto.request.SignupRequestDto;
 import com.sns.api.auth.service.AuthService;
 import com.sns.api.common.domain.dto.UserBaseDto;
-import com.sns.api.users.domain.dto.UsersResponseDto;
+import com.sns.api.users.domain.dto.response.UsersResponseDto;
 import com.sns.common.component.BaseResponse;
 import com.sns.common.component.Const;
 import com.sns.common.component.ResultCode;
