+++ src/main/java/com/sns/api/auth/domain/dto/request/LoginRequestDto.java
@@ -1,11 +1,14 @@
-package com.sns.api.users.domain.dto;
+package com.sns.api.auth.domain.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 
-
 @Getter
-public class UserDeleteRequestDto{
+public class LoginRequestDto {
+
+    @NotBlank(message = "이메일을 입력해주세요.")
+    private String email;
+
     @NotBlank(message = "비밀번호를 입력해주세요.")
     private String password;
 }
