+++ src/main/java/com/sns/api/users/domain/dto/request/UserDeleteRequestDto.java
@@ -1,14 +1,11 @@
-package com.sns.api.auth.domain.dto;
+package com.sns.api.users.domain.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 
+
 @Getter
-public class LoginRequestDto {
-
-    @NotBlank(message = "이메일을 입력해주세요.")
-    private String email;
-
+public class UserDeleteRequestDto{
     @NotBlank(message = "비밀번호를 입력해주세요.")
     private String password;
 }
