+++ src/main/java/com/sns/api/users/service/impl/UsersServiceImpl.java
@@ -1,11 +1,11 @@
 package com.sns.api.users.service.impl;
 
 import com.sns.api.friends.repository.FriendsRepository;
-import com.sns.api.users.domain.dto.UserReadResponseDto;
-import com.sns.api.users.domain.dto.UserUpdateRequestDto;
-import com.sns.api.users.domain.dto.PasswordUpdateDto;
-import com.sns.api.users.domain.dto.UserDeleteRequestDto;
-import com.sns.api.users.domain.dto.UsersResponseDto;
+import com.sns.api.users.domain.dto.response.UserReadResponseDto;
+import com.sns.api.users.domain.dto.request.UserUpdateRequestDto;
+import com.sns.api.users.domain.dto.request.PasswordUpdateDto;
+import com.sns.api.users.domain.dto.request.UserDeleteRequestDto;
+import com.sns.api.users.domain.dto.response.UsersResponseDto;
 import com.sns.api.users.domain.entity.Users;
 
 import com.sns.api.users.repository.UsersRepository;
